cambios de decimales y anexo de vendedor en el ticket

+++ database/migrations/2020_06_16_000423_create_provider_purchase_order_details_table.php
@@ -26,13 +26,13 @@ class CreateProviderPurchaseOrderDetailsTable extends Migration
             $table->bigInteger('meas_fk')->unsigned(); //Llave Foranea Catálogo Unidad de Medida
             $table->foreign('meas_fk')->references('meas_pk')->on('measurements')->onUpdate('cascade');
 
-            $table->integer('ppod_quantity'); //Cantidad
+            $table->decimal('ppod_quantity', 12, 2); //Cantidad
             $table->decimal('ppod_actualprice', 12, 2); //Precio Actual
             $table->decimal('ppod_providerprice', 12, 2); //Precio Proveedor
             $table->decimal('ppod_discountrate', 12, 2); //Porcentaje Descuento
             $table->decimal('ppod_ieps', 12, 2); //IEPS
             $table->decimal('ppod_iva', 12, 2); //IVA
-            $table->smallInteger('ppod_status')->default(1); //Estatus
+            $table->smallInteger('ppod_status')->default(1); //Estatus 
             $table->timestamps();
         });
     }
